lambdas/qr_scanner/requirements.txt
# boto3
Pillow
opencv-python-headless
pyzbar
numpy
